requirements.txt
apache-airflow==2.7.3
pandas==2.1.0
requests==2.31.0
transformers==4.35.0
torch==2.1.0
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.21
python-dotenv==1.0.0
pytest==7.4.3
feedparser==6.0.10
